Passenger Logic
Created a passenger prefab which holds the passenger script. This script handles behaviors like being picked up and dropped off. Created a script called Passenger Interactor which goes on the player and handles storing the players and telling them when they are picked up and dropped off. There is also a player spawner which spawns a player randomly every x seconds, and a passenger data script which stores data like original and destination locations.